app_flask/requirements.txt
aniso8601==9.0.1
astroid==3.2.2
black==24.4.2
blinker==1.7.0
click==8.1.7
dill==0.3.8
Flask==3.0.2
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pylint==3.2.2
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.28
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.10.0
Werkzeug==3.0.1
